Adding product, Edit business information, display products on business fragment
Edit Business Fragement
Business List adapter
Business list item view holder

+++ app/src/main/java/com/example/shopit/data/cart/CartProductDataClass.kt
@@ -3,7 +3,7 @@ package com.example.shopit.data.cart
 data class CartProductDataClass (
     var cartProductImage: String,   //URL
     var cartProductName: String,    //Name
-    var cartProductPrice: Float,   //Price
+    var cartProductPrice: Double,   //Price
     var cartProductQuantity: Int,
     var cartProductBarcode: String  //Barcode numbers
     )
